requirements.txt
requests
pandas
mysql-connector-python
mplfinance
matplotlib
python-dotenv
flask
apscheduler
flask-cors
